Decimation is now a legacy feature
